feat: policy for registrations

+++ src/Controller/RegistrationsController.php
@@ -3,6 +3,10 @@ declare(strict_types=1);
 
 namespace App\Controller;
 
+use App\Model\Entity\Registration;
+use Authorization\Exception\ForbiddenException;
+use InvalidArgumentException;
+
 /**
  * Registrations Controller
  *
@@ -18,10 +22,16 @@ class RegistrationsController extends AppController
      */
     public function index($editionID = null)
     {
+        $this->Authorization->skipAuthorization();
         $edition = $this->getTableLocator()->get('Editions');
-        $editionName = $edition->findById($editionID)->firstOrFail()->title;
-        $registrations = $this->paginate($this->Registrations->findByEditionid($editionID));
 
+        if($this->getUser()->isAdmin){
+            $registrations = $this->paginate($this->Registrations->findByEditionid($editionID));
+        } else {
+            $registrations = $this->paginate($this->Registrations->findByUserid($this->getUser()->id));
+        }
+        
+        $editionName = $edition->findById($editionID)->firstOrFail()->title;
         $this->set(compact('registrations'));
         $this->set(compact('editionID'));
         $this->set(compact('editionName'));
@@ -39,11 +49,20 @@ class RegistrationsController extends AppController
         $registration = $this->Registrations->get($id, [
             'contain' => [],
         ]);
+        $this->authorize($registration);
 
         $edition = $this->getTableLocator()->get('Editions');
         $editionName = $edition->findById($editionID)->firstOrFail()->title;
 
+        $schools = $this->getTableLocator()->get('Schools');
+        $school = $schools->findById($registration->schoolId)->firstOrFail();
+
+        $students = $this->getTableLocator()->get('Students');
+        $students = $this->paginate($students->findBySchoolid($registration->schoolId));
+
         $this->set(compact('editionID'));
+        $this->set(compact('students'));
+        $this->set(compact('school'));
         $this->set(compact('editionName'));
         $this->set(compact('registration'));
     }
@@ -58,20 +77,34 @@ class RegistrationsController extends AppController
 
         $edition = $this->getTableLocator()->get('Editions');
         $editionName = $edition->findById($editionID)->firstOrFail()->title;
+
+        $schools = $this->getTableLocator()->get('Schools');
+        $school = $schools->findByUserid($this->getUser()->id)->firstOrFail();
+
+        $students = $this->getTableLocator()->get('Students');
+        $students = $students->findBySchoolid($school->id);
+
         $registration = $this->Registrations->newEmptyEntity();
+
+        $this->authorize($registration);
+
         if ($this->request->is('post')) {
 
-            var_dump($this->request->getData('students'));
-            var_dump($this->request->getData('schools'));
             $registration = $this->Registrations->patchEntity($registration, $this->request->getData());
-            if ($this->Registrations->save($registration)) {
-                $this->Flash->success(__('The registration has been saved.'));
+            
+            $registration->userId = $this->getUser()->id;
+            $registration->schoolId = $school->id;
+            $registration->editionId = $editionID;
 
+            if ($this->Registrations->save($registration)) {
+                $this->updateStudent($this->request->getData('students'),$registration->id);
+                $this->Flash->success(__('The registration has been saved.'));
                 return $this->redirect(['action' => 'index', $editionID]);
             }
             $this->Flash->error(__('The registration could not be saved. Please, try again.'));
         }
         $this->set(compact('editionID'));
+        $this->set(compact('students'));
         $this->set(compact('editionName'));
         $this->set(compact('registration'));
     }
@@ -90,20 +123,51 @@ class RegistrationsController extends AppController
         $registration = $this->Registrations->get($id, [
             'contain' => [],
         ]);
+        $schools = $this->getTableLocator()->get('Schools');
+        $school = $schools->findById($registration->schoolId)->firstOrFail();
+
+        $students = $this->getTableLocator()->get('Students');
+        $students = $students->findBySchoolid($registration->schoolId);
+
+        $this->authorize($registration);
         if ($this->request->is(['patch', 'post', 'put'])) {
             $registration = $this->Registrations->patchEntity($registration, $this->request->getData());
+            $registration->userId = $this->getUser()->id;
             if ($this->Registrations->save($registration)) {
+                if(!$this->getUser()->isAdmin && !empty($this->request->getData('students'))){
+                    $this->updateStudent($this->request->getData('students'),$registration->id);
+                }
                 $this->Flash->success(__('The registration has been saved.'));
 
                 return $this->redirect(['action' => 'index', $editionID]);
             }
             $this->Flash->error(__('The registration could not be saved. Please, try again.'));
         }
+        
+        $isAdmin = $this->getUser()->isAdmin;
+        $this->set(compact('isAdmin'));
         $this->set(compact('editionID'));
+        $this->set(compact('students'));
         $this->set(compact('editionName'));
         $this->set(compact('registration'));
     }
 
+    private function updateStudent($studentsIdSelected, $registrationId){
+
+        $schools = $this->getTableLocator()->get('Schools');
+        $school = $schools->findByUserid($this->getUser()->id)->firstOrFail();
+
+        $table = $this->getTableLocator()->get('Students');
+        $students = $table->findBySchoolid($school->id);
+
+        foreach($students as $student){
+            foreach($studentsIdSelected as $idStudentSelected){
+                if($student->id == $idStudentSelected){
+                    $table->updateAll(['registrationId' => $registrationId], ['id' => $student->id]);
+                }
+            }
+        }
+    }
     /**
      * Delete method
      *
@@ -115,6 +179,7 @@ class RegistrationsController extends AppController
     {
         $this->request->allowMethod(['post', 'delete']);
         $registration = $this->Registrations->get($id);
+        $this->authorize($registration);
         if ($this->Registrations->delete($registration)) {
             $this->Flash->success(__('The registration has been deleted.'));
         } else {
@@ -122,5 +187,18 @@ class RegistrationsController extends AppController
         }
 
         return $this->redirect(['action' => 'index', $editionID]);
+    }
+
+    private function authorize(Registration $r){
+        try{
+            $this->Authorization->authorize($r);
+        } catch(ForbiddenException $e){
+            $this->Flash->error("Vous n'avez pas l'autorisation.");
+            return $this->redirect(['controller' => 'Registrations', 'action' => 'index']);
+        } 
+    }
+
+    private function getUser(){
+        return $this->Authentication->getResult()->getData();
     }
 }
